test-requirements.txt
hacking
mock
ansible-lint==4.2.0
docker==4.1.0
ansible-test
pytest
python-vagrant==0.5.15
